move threads code to separated file
The thread code will be improved to support dynamic tasks.
